+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Changelog
 * Update :command:`run-ngs-pipeline` and :command:`run-chip-pipeline` commands to be able to subset/exclude specified samples.
 * Remove ``--fn`` argument from :command:`filter-samples` command.
 * Update CNV caller for CYP2D6.
+* Update :meth:`api.plot.plot_cn_af` method to accept both VcfFrame[Imported] and VcfFrame[Consolidated].
 
 0.9.0 (2021-12-05)
 ------------------